ecommerce_backend/requirements.txt
asgiref==3.9.1
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==25.0
PyJWT==2.10.1
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
